requirements.txt
dj-database-url==0.5.0
Django==2.2.23
django-filter==2.4.0
django-webpack-loader==1.0.0
djangorestframework==3.11.2
gunicorn==20.1.0
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
psycopg2==2.8.6


